(DOCSP-42870) LangChain JS/TS: Create a Vector Search index procedures (#6106)
* remove procedure and add code block to make index

* change code block language

* add run command

* update dependencies & imports

* empty collection if filled, create index only if not yet created

* .

* empty collection if filled

* create index only if not yet created

+++ source/includes/ai-integrations/langchain/langchain-create-vector-store.js
@@ -11,7 +11,10 @@ async function run() {
     };
     
     // Ensure that the collection is empty
-    await collection.deleteMany({});
+    const count = await collection.countDocuments();
+    if (count > 0) {
+      await collection.deleteMany({});
+    }
 
     // Save online PDF as a file
     const rawData = await fetch("https://query.prod.cms.rt.microsoft.com/cms/api/am/binary/RE4HkJP");
